+++ pkg/llm-d-inference-sim/response.go
@@ -24,6 +24,9 @@ import (
 	"strings"
 )
 
+// completionResponse interface representing both completion response types (text and chat)
+type completionResponse interface{}
+
 // baseCompletionResponse contains base completion response related information
 type baseCompletionResponse struct {
 	// ID defines the response ID
@@ -48,16 +51,19 @@ type usage struct {
 	TotalTokens int `json:"total_tokens"`
 }
 
+// chatCompletionResponse defines structure of /chat/completion response
+type chatCompletionResponse struct {
+	baseCompletionResponse
+	// Choices list of Choices of the response, according of OpenAI API
+	Choices []chatRespChoice `json:"choices"`
+}
+
 // baseResponseChoice contains base completion response's choice related information
 type baseResponseChoice struct {
 	// Index defines completion response choise Index
 	Index int `json:"index"`
 	// FinishReason defines finish reason for response or for chunks, for not last chinks is defined as null
 	FinishReason *string `json:"finish_reason"`
-}
-
-// completionResponse interface representing both completion response types (text and chat)
-type completionResponse interface {
 }
 
 // v1/chat/completion
@@ -147,13 +153,6 @@ type toolCall struct {
 	Type string `json:"type"`
 	// Index is the index of the tool in the sequence of tools generated by the model
 	Index int `json:"index"`
-}
-
-// chatCompletionResponse defines structure of /chat/completion response
-type chatCompletionResponse struct {
-	baseCompletionResponse
-	// Choices list of Choices of the response, according of OpenAI API
-	Choices []chatRespChoice `json:"choices"`
 }
 
 // chatRespChoice represents a single chat completion response choise
@@ -292,7 +291,7 @@ func (req textCompletionRequest) createToolCalls() ([]toolCall, string, error) {
 	return nil, "", errors.New("tool calls are not supported in text completion")
 }
 
-// completionError defines structure of error returned by completion request
+// completionError defines the simulator's response in case of an error
 type completionError struct {
 	// Object is a type of this Object, "error"
 	Object string `json:"object"`
