+++ pkg/llm-d-inference-sim/simulator.go
@@ -33,13 +33,74 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/google/uuid"
 	vllmapi "github.com/llm-d/llm-d-inference-sim/pkg/vllm-api"
+	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/spf13/pflag"
 	"github.com/valyala/fasthttp"
 	"github.com/valyala/fasthttp/fasthttpadaptor"
 )
 
-const vLLMDefaultPort = 8000
+const (
+	vLLMDefaultPort           = 8000
+	modeRandom                = "random"
+	modeEcho                  = "echo"
+	chatComplIDPrefix         = "chatcmpl-"
+	stopFinishReason          = "stop"
+	lengthFinishReason        = "length"
+	toolsFinishReason         = "tool_calls"
+	roleAssistant             = "assistant"
+	roleUser                  = "user"
+	textCompletionObject      = "text_completion"
+	chatCompletionObject      = "chat.completion"
+	chatCompletionChunkObject = "chat.completion.chunk"
+	toolChoiceNone            = "none"
+	toolChoiceAuto            = "auto"
+	toolChoiceRequired        = "required"
+)
+
+// VllmSimulator simulates vLLM server supporting OpenAI API
+type VllmSimulator struct {
+	// logger is used for information and errors logging
+	logger logr.Logger
+	// timeToFirstToken time before the first token will be returned, in milliseconds
+	timeToFirstToken int
+	// interTokenLatency time between generated tokens, in milliseconds
+	interTokenLatency int
+	// port defines on which port the simulator runs
+	port int
+	// mode defines the simulator response generation mode, valid values: echo, random
+	mode string
+	// model defines the current base model name
+	model string
+	// loraAdaptors contains list of LoRA available adaptors
+	loraAdaptors sync.Map
+	// maxLoras defines maximum number of loaded loras
+	maxLoras int
+	// maxCPULoras defines maximum number of loras to store in CPU memory
+	maxCPULoras int
+	// runningLoras is a collection of running loras, key of lora's name, value is number of requests using this lora
+	runningLoras sync.Map
+	// waitingLoras will represent collection of loras defined in requests in the queue - Not implemented yet
+	waitingLoras sync.Map
+	// maxRunningReqs defines the maximum number of inference requests that could be processed at the same time
+	maxRunningReqs int64
+	// nRunningReqs is the number of inference requests that are currently being processed
+	nRunningReqs int64
+	// nWaitingReqs is the number of inference requests that are waiting to be processed
+	nWaitingReqs int64
+	// loraInfo is prometheus gauge
+	loraInfo *prometheus.GaugeVec
+	// runningRequests is prometheus gauge
+	runningRequests *prometheus.GaugeVec
+	// waitingRequests is prometheus gauge for number of queued requests
+	waitingRequests *prometheus.GaugeVec
+	// kvCacheUsagePercentage is prometheus gauge
+	kvCacheUsagePercentage *prometheus.GaugeVec
+	// channel for requeasts to be passed to workers
+	reqChan chan *completionReqCtx
+	// schema validator for tools parameters
+	toolsValidator *validator
+}
 
 // New creates a new VllmSimulator instance with the given logger
 func New(logger logr.Logger) (*VllmSimulator, error) {
@@ -230,30 +291,12 @@ func (s *VllmSimulator) HandleTextCompletions(ctx *fasthttp.RequestCtx) {
 
 func (s *VllmSimulator) HandleLoadLora(ctx *fasthttp.RequestCtx) {
 	s.logger.Info("load lora request received")
-
-	var req loadLoraRequest
-	err := json.Unmarshal(ctx.Request.Body(), &req)
-	if err != nil {
-		s.logger.Error(err, "failed to read and parse load lora request body")
-		ctx.Error("failed to read and parse load lora request body, "+err.Error(), fasthttp.StatusBadRequest)
-		return
-	}
-
-	s.addLora(req.LoraName)
+	s.loadLora(ctx)
 }
 
 func (s *VllmSimulator) HandleUnloadLora(ctx *fasthttp.RequestCtx) {
 	s.logger.Info("load lora request received")
-
-	var req unloadLoraRequest
-	err := json.Unmarshal(ctx.Request.Body(), &req)
-	if err != nil {
-		s.logger.Error(err, "failed to read and parse unload lora request body")
-		ctx.Error("failed to read and parse unload lora request body, "+err.Error(), fasthttp.StatusBadRequest)
-		return
-	}
-
-	s.removeLora(req.LoraName)
+	s.unloadLora(ctx)
 }
 
 // isValidModel checks if the given model is the base model or one of "loaded" LoRAs
